commit version with double before reverting it to int

+++ includes/wolf3d.h
@@ -6,7 +6,7 @@
 /*   By: cheron <cheron@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/01/19 14:40:51 by cheron            #+#    #+#             */
-/*   Updated: 2015/01/23 18:33:41 by cheron           ###   ########.fr       */
+/*   Updated: 2015/01/26 16:29:53 by cheron           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,8 +39,8 @@ typedef	struct		s_map
 
 typedef struct		s_coord
 {
-	int				x;
-	int				y;
+	double			x;
+	double			y;
 }					t_coord;
 
 typedef	struct		s_pdata
@@ -63,8 +63,8 @@ t_pdata				*ft_allocpdata(void);
 void				ft_freepdata(t_pdata *pdata);
 void				ft_print_map(t_map *map);
 void				set_player_pos(t_map *map, t_coord *pcoord);
-int					inter_vertical(t_map *map, t_pdata *pdata, double alpha);
-int					inter_horizontal(t_map *map, t_pdata *pdata, double alpha);
+double				inter_vertical(t_map *map, t_pdata *pdata, double alpha);
+double				inter_horizontal(t_map *map, t_pdata *pdata, double alpha);
 SDL_Window			*ft_createwin(char *title, int height, int width);
 void				ft_proceed(t_screen *screen, t_map *map, t_pdata *pdata);
 void				ft_DrawRenderer(t_pdata *pdat, t_map *map, SDL_Renderer *r);
